refactor(projects): refactored
